Split up menu/handlers/and commands
Made the code more readable, handlers for init, config, and token has been seperated. CLI menu variants have also been seperated
