+++ commands/confighelp.txt
@@ -1,0 +1,10 @@
+
+myapp <command> <option>
+
+Usage:
+
+myapp config --show     displays a list of the current config settings
+myapp config --reset    resets the config file with default settings
+myapp config --set      sets a specific config setting
+myapp config --help
+
