+++ commands/inithelp.txt
@@ -1,0 +1,9 @@
+
+myapp init <command> <option>
+
+Usage:
+
+myapp init --all          creates the folder structure and config file
+myapp init --mk           creates the folder structure
+myapp init --cat          creates the config file with default settings
+myapp init --help
